+++ CHANGELOG.md
@@ -1,3 +1,15 @@
-## [0.0.1] - TODO: Add release date.
+# Changelog
 
-* TODO: Describe initial release.
+## [1.1.0]
+
+- Updated `pubspec.yaml`
+- Added example application called example
+- Updated the way the mime types are accessed
+- Changed `createFile` function to include preliminary data
+- Added export mime type option to `appendFile`
+- Added command to open file from fileID
+- Added `getData` to get data as plain text
+
+## [1.0.0]
+
+- Initial commit